s_shiftRightJamM: avoid accidental corruption when size_words = 0
More observant compilers complain there is a potential path through
the code that leaves ptr undefined. This would occur in the unlikely
but possible event that size_words was 0 while dist>>5 was also zero.

We can fix this by doing the initial test for wordDist before the
first iteration of 0 termination.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>